+++ src/ast/semantic_analyser.h
@@ -106,6 +106,7 @@ private:
   ProbeType single_provider_type(void);
   AddrSpace find_addrspace(ProbeType pt);
 
+  void binop_ptr(Binop &op);
   void binop_int(Binop &op);
 
   bool in_loop(void)