+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to
   - [#1900](https://github.com/iovisor/bpftrace/pull/1900)
 - Support for C style integer suffix in parser
   - [#1938](https://github.com/iovisor/bpftrace/pull/1938)
+- Add C like pointer arithmetic
+  - [#1881](https://github.com/iovisor/bpftrace/pull/1881)
 
 #### Changed